[WasmFS] Add a hack to work around deadlocks
When the main thread tries to perform a WasmFS operation that tries to acquire a
lock that is already held by another thread that is trying to proxy a backend
operation, deadlock can occur. Proxying depends on passing a message to a worker
thread and that message has to be relayed through the main thread. When the main
thread is spinning on a lock, the proxying message is not relayed and neither
thread makes forward progress.

Work around the issue by adding a heartbeat event on the dedicated worker thread
to make it execute the queue periodically even without receiving a message from
the main thread. This allows the proxying thread holding the lock to make
forward progress and resolves the deadlock.